+++ admin/friendly_urls/friendly_url_manager.php
@@ -4,53 +4,93 @@
     require_once CMS_ROOT . 'database/dao/friendly_url_dao.php';
     require_once CMS_ROOT . 'database/dao/settings_dao.php';
     require_once CMS_ROOT . 'database/dao/page_dao.php';
+    require_once CMS_ROOT . 'database/dao/article_dao.php';
 
     class FriendlyUrlManager {
 
         private $_friendly_url_dao;
         private $_settings_dao;
         private $_page_dao;
+        private $_article_dao;
 
         public function __construct() {
             $this->_friendly_url_dao = FriendlyUrlDao::getInstance();
             $this->_settings_dao = SettingsDao::getInstance();
             $this->_page_dao = PageDao::getInstance();
+            $this->_article_dao = ArticleDao::getInstance();
             $this->writeHtaccessFileIfNotExists();
         }
 
         public function insertOrUpdateFriendlyUrlForPage($page) {
-            $url = '/' . $this->createNewUrlFor($page);
-            $url = $this->appendNumberIfFriendlyUrlExists($url, $page);
-            $this->_friendly_url_dao->insertOrUpdateFriendlyUrl($url, $page);
+            $url = '/' . $this->createUrlForPage($page);
+            $this->insertOrUpateFriendlyUrl($url, $page);
+        }
+
+        public function insertOrUpdateFriendlyUrlForArticle($article) {
+            $url = '/' . $this->createUrlForArticle($article);
+            $this->insertOrUpateFriendlyUrl($url, $article);
         }
 
         public function getPageFromUrl($url) {
-            $url = rtrim($url, '/');
-            return $this->_friendly_url_dao->getPageFromUrl($url);
+            $url = $this->removeQueryStringFrom($url);
+            $element_holder_id = $this->_friendly_url_dao->getElementHolderIdFromUrl($url);
+            $page = $this->_page_dao->getPageByElementHolderId($element_holder_id);
+            if (is_null($page)) {
+                $page_part_of_url = substr($url, 0, strrpos($url, '/'));
+                $element_holder_id = $this->_friendly_url_dao->getElementHolderIdFromUrl($page_part_of_url);
+                $page = $this->_page_dao->getPageByElementHolderId($element_holder_id);
+            }
+            return $page;
         }
 
-        public function getFriendlyUrlForPage($page) {
-            return $this->_friendly_url_dao->getUrlFromPage($page);
+        public function getArticleFromUrl($url) {
+            $url = $this->removeQueryStringFrom($url);
+            $element_holder = null;
+            $url_parts = explode('/', $url);
+            if (count($url_parts) > 1) {
+                $element_holder_id = $this->_friendly_url_dao->getElementHolderIdFromUrl(substr($url, strrpos($url, '/'), strlen($url) - 1));
+                return $this->_article_dao->getArticleByElementHolderId($element_holder_id);
+            }
         }
 
-        private function createNewUrlFor($page) {
+        public function getFriendlyUrlForElementHolder($element_holder) {
+            return $this->_friendly_url_dao->getUrlFromElementHolder($element_holder);
+        }
+
+        private function insertOrUpateFriendlyUrl($url, $element_holder) {
+            $url = $this->appendNumberIfFriendlyUrlExists($url, $element_holder);
+            if (is_null($this->getFriendlyUrlForElementHolder($element_holder)))
+                $this->_friendly_url_dao->insertFriendlyUrl($url, $element_holder);
+            else
+                $this->_friendly_url_dao->updateFriendlyUrl($url, $element_holder);
+        }
+
+        private function createUrlForPage($page) {
             $url = $this->replaceSpecialCharacters($page->getNavigationTitle());
             $parent_page = $page->getParent();
             if ($parent_page != null && $parent_page->getId() != $this->_page_dao->getRootPage()->getId())
-                $url = $this->createNewUrlFor($page->getParent()) . "/" . $url;
+                $url = $this->createNewUrlForPage($page->getParent()) . "/" . $url;
             return $url;
         }
 
-        private function appendNumberIfFriendlyUrlExists($url, $page) {
+        private function createUrlForArticle($article) {
+            return $this->replaceSpecialCharacters($article->getTitle());
+        }
+
+        private function appendNumberIfFriendlyUrlExists($url, $element_holder) {
             $new_url = $url;
-            $existing_page = $this->getPageFromUrl($url);
+            $existing_element_holder_id = $this->_friendly_url_dao->getElementHolderIdFromUrl($url);
             $number = 1;
-            while ($existing_page != null && $existing_page->getId() != $page->getId()) {
+            while ($existing_element_holder_id != null && $existing_element_holder_id != $element_holder->getId()) {
                 $new_url = $url . $number;
                 $number++;
-                $existing_page = $this->getPageFromUrl($new_url);
+                $existing_element_holder_id = $this->_friendly_url_dao->getElementHolderIdFromUrl($new_url);
             }
             return $new_url;
+        }
+
+        private function removeQueryStringFrom($url) {
+            return strtok(rtrim($url, '/'), '?');
         }
 
         private function replaceSpecialCharacters($value) {