+++ admin/utilities/url_helper.php
@@ -1,0 +1,13 @@
+<?php
+    defined('_ACCESS') or die;
+
+    class UrlHelper {
+
+        public static function addQueryStringParameter($url, $name, $value) {
+            if (strpos('?', $url) !== false)
+                return $url . '&' . $name . '=' . $value;
+            else
+                return $url . '?' . $name . '=' . $value;
+        }
+
+    }